Fix race condition when resizing buffer during borrows.
We don't resize the buffer if the requested size is < current size,
but it's possible someone else is resizing the buffer when we check
that, so we need to check it again once we hold the lock to make
sure we're not shrinking the buffer.

We also need to poll when waiting for write space to catch when the
buffer expands in size.